# Fix `asv` speed benchmark