--- a/asv.conf.json
+++ b/asv.conf.json
@@ -71,7 +71,9 @@
     // pip (with all the conda available packages installed first,
     // followed by the pip installed packages).
     //
-    "matrix": {},
+    "matrix": {
+        "fakeredis": [""],
+    },
 
     // Combinations of libraries/python versions can be excluded/included
     // from the set to test. Each entry is a dictionary containing additional